% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_cdf.R
\name{duos_cdf}
\alias{duos_cdf}
\title{Bayesina Density Estimate of CDF from DUOS.}
\usage{
duos_cdf(x, duos_output, burnin = NA, scale = FALSE)
}
\arguments{
\item{x}{A single value or vector of values to calculate the CDF at}

\item{duos_output}{The list returned by \code{duos}}

\item{burnin}{The desired burnin to discard from including in the estimate}
}
\description{
Calculates an estimate of the CDF based on the output from \code{duos} for an individual
or vector of values
}
