% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_cdf_plot.R
\name{duos_cdf_plot}
\alias{duos_cdf_plot}
\title{Plot CDF from DUOS.}
\usage{
duos_cdf_plot(duos_output, burnin = NA, cr_i = FALSE, data = FALSE)
}
\arguments{
\item{duos_output}{The list returned by \code{duos}}

\item{burnin}{The desired burnin to discard from including in the estimate}

\item{cr_i}{An option to include credible intervals}

\item{data}{An option to overlay empirical CDF}
}
\description{
Plots the CDF estimate based on the output from DUOS.
}
