% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_mcmcplots.R
\name{duos_mcmcplots}
\alias{duos_mcmcplots}
\title{Plots a Variety of Convergence Diagnostic Plots}
\usage{
duos_mcmcplots(duos_output, type = "traceplot", parameters = "c", plots = "all", burnin = 1)
}
\arguments{
\item{duos_output}{The list returned by \code{duos} containing the density estimate results.}

\item{type}{The type of convergent plot to create (see details).}

\item{parameters}{The group of parameters to plot (see details).}

\item{plots}{An option to plot parameters on the same plot or as individuals in a grid (see details).}

\item{burnin}{The desired burnin to discard from the results. By default, it is 1 so that all iterations are plotted.}
}
\value{
A plot of trace plots.
}
\description{
Plots the convergence plots on the parameters from \code{duos}.
}
\details{
\strong{Options for} \code{parameters}

There are two sets of parameters that can plotted in the trace plots: the cut-points and the bin proportion parameters.
\itemize{
    \item \code{"c"}: Plots the trace plots of the cut-points that are ordered and between 0 and 1 (DEFAULT).
    \item \code{"p"}: Plots the trace plots of the proportion parameters that sum to one.
  }

\strong{Options for} \code{type}

There are several options on which convergence plots to create.
\itemize{
    \item \code{"traceplot"}: Creates trace plots on the parameters (DEFAULT).
    \item \code{"acf"}: Creates autocorrelation plots using lag=1000. The burnin should be set to the desired value to discard. 
    \item \code{"rm"}: Creates running mean plots on the parameters.
  }
  
\strong{Options for} \code{plots}

There are several options on how to display the trace plots.
\itemize{
    \item \code{"all"}: Overlays all trace plots in one plot (works well for the cut-point parameters) (DEFAULT).
    \item \code{"indiv"}: Create a grid of trace plots where each plot contains a single parameter's trace plot. Six plots are allowed in a grid so multiple graphs are created if there are more than 6 parameters.
  }
}
\examples{

## --------------------------------------------------------------------------------
## Uniform Distribution
## --------------------------------------------------------------------------------

# First run 'duos' on data sampled from a Uniform(0,1) distribution with 50 data points.
y <- runif(50)
duos_unif <- duos(y = y)

# Plot the trace plots of the cut-points on a single graph (trace plot is the default)
duos_mcmcplots(duos_unif)

# Plot the acf plots of the cut-points on separate graphs
duos_mcmcplots(duos_unif, type = "acf", plots = "indiv", burnin = 10000)

## --------------------------------------------------------------------------------
## Beta Distribution
## --------------------------------------------------------------------------------

# First run 'duos' on data sampled from a Beta(0.5,0.5) distribution with 300 data points.
y <- rbeta(300, 0.5, 0.5)
duos_arcsin <- duos(y = y, k = 10, MH_N = 20000)

#Plot the trace plots for the cut-points as individual graphs
#Note: The plots are printed six at a time so multiple panels are printed
duos_mcmcplots(duos_arcsin, plots = "indiv")

#Plot the trace plots for the bin proportions as individual graphs
#Note: The plots are printed six at a time so multiple panels are printed
duos_mcmcplots(duos_arcsin, parameters = "p", plots = "indiv")

## --------------------------------------------------------------------------------
## Bimodal Distribution
## --------------------------------------------------------------------------------

# Sample 150 random uniforms
u <- runif(150)
y <- rep(NA, 150)
# Sampling from the mixture
for(i in 1:150){
  if(u[i]<.3){
   y[i] <- rnorm(1, 0, 1)
  }else {
   y[i] <- rnorm(1, 4, 1)
  }
}
# First run 'duos' on data sampled from a bimodal distribution with 150 data points.
duos_bimodal <- duos(y = y, k = 8, MH_N = 20000)

# Plot the running mean plots for the cut-point parameters 
duos_mcmcplots(duos_bimodal, type = "rm", parameters = "c")

# Plot the autocorrelation plots for the bin proportions with a burnin of 10,000
duos_mcmcplots(duos_bimodal, type = "acf", parameters = "p", burnin = 10000)
}
