% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_pdf.R
\name{duos_pdf}
\alias{duos_pdf}
\title{Bayesina Density Estimate of PDF from DUOS.}
\usage{
duos_pdf(x, duos_output, burnin)
}
\arguments{
\item{x}{A single value or vector of values to calculate the density at}

\item{duos_output}{The list returned by \code{duos}}

\item{burnin}{The desired burnin to discard from including in the estimate}
}
\description{
Calculates an estimate of the density based on the output from \code{duos} for an individual
or vector of values
}
