% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_pdf_plot.R
\name{duos_pdf_plot}
\alias{duos_pdf_plot}
\title{Plot Probability Density from DUOS.}
\usage{
duos_pdf_plot(duos_output, burnin = NA, cr_i = FALSE, data = FALSE)
}
\arguments{
\item{duos_output}{The list returned by \code{duos}}

\item{burnin}{The desired burnin to discard from including in the estimate}

\item{cr_i}{An option to include credible intervals}

\item{data}{An option to overlay histogram of original data on plot}
}
\description{
Plots the density estimate based on the output from DUOS.
}
