% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_plot.R
\name{duos_plot}
\alias{duos_plot}
\title{Plot PDF or CDF}
\usage{
duos_plot(duos_output, type = "pdf", burnin = NA, cri = FALSE,
  data = FALSE)
}
\arguments{
\item{duos_output}{The list returned by \code{duos} containing the density estimate results.}

\item{type}{The type of desired graph (see details).}

\item{burnin}{The desired burnin to discard from the results. If no values is entered, the default is half the number of iterations.}

\item{cri}{An option to include credible intervals.}

\item{data}{An option to include data in the graph (see details).}
}
\value{
A plot of the PDF or CDF estimate.
}
\description{
Plots the posterior mean PDF or CDF based on the output from \code{duos}.
}
\details{
The form of the density whose paramters are estimated in \code{duos} is below:

\deqn{f(x) =}

\deqn{(\pi_1) / (\gamma_1) , 0 \le x < \gamma_1}
\deqn{(\pi_2) / (\gamma_2-\gamma_1) , \gamma_1 \le x  < \gamma_2}
\deqn{(\pi_3) / (\gamma_3-\gamma_2) , \gamma_2 \le x  < \gamma_3}
\deqn{...  , ... \le x  < ...}
\deqn{(\pi_k) / (\gamma_k-\gamma_{k-1}) , \gamma_{k-1} \le x  < \gamma_k}
\deqn{(\pi_{k+1}) / (1-\gamma_k) , \gamma_k \le x  < 1}

\deqn{F(x) =}

\deqn{(\pi_1) / (\gamma_1) * x , 0 \le x < \gamma_1}
\deqn{\pi_1 + (\pi_2) / (\gamma_2-\gamma_1) * (x-\gamma_1) , \gamma_1 \le x  < \gamma_2}
\deqn{\pi_1 + \pi_2 + (\pi_3) / (\gamma_3-\gamma_2) * (x-\gamma_2) , \gamma_2 \le x  < \gamma_3}
\deqn{...  , ... \le x  < ...}
\deqn{\sum_{i=1}^{k-1} \pi_i + (\pi_k) / (\gamma_k-\gamma_{k-1}) * (x-\gamma_{k-1}) , \gamma_{k-1} \le x  < \gamma_k}
\deqn{\sum_{i=1}^{k}\pi_i + (\pi_{k+1}) / (1-\gamma_k) * (x-\gamma_k) , \gamma_k \le x  < 1}

\strong{Options for} \code{type}

The input from \code{duos} can be used to plot a Bayesian estimate of the PDF or the CDF.
\itemize{
    \item \code{"pdf"}: The density, f(x), is calculated at a grid of 999 points between 0 and 1 at each iteration. The density at each grid point is then averaged across the iterations to produce the results in the graph (DEFAULT).
    \item \code{"cdf"}: The CDF, F(x), is calculated at a grid of 999 points between 0 and 1 at each iteration. The CDF at each grid point is then averaged across the iterations to produce the results in the graph.
  }

\strong{Options for} \code{cri}

Credibal intervals can also be added to the plot of the PDF or CDF.
\itemize{
    \item \code{"FALSE"}: No credible intervals lines are plotted (DEFAULT).
    \item \code{"TRUE"}: Credible interval lines are plotted in red on the PDF or CDF. These are calculated by taking the 0.025th and 0.975th quantiles of the iterations from \code{duos} after burnin.
  }

\strong{Options for} \code{data}

Incoproates the data into the PDF and CDF.
\itemize{
    \item \code{"FALSE"}: The data is not included (DEFAULT).
    \item \code{"TRUE"}: If the PDF is plotted, a histogram is overlayed with the density estimate. If the CDF is plotted, the empirical CDF is overlayed with the \code{duos} CDF estimate.
  }
}
\examples{

## --------------------------------------------------------------------------------
## Uniform Distribution
## --------------------------------------------------------------------------------

# First run 'duos' on data sampled from a Uniform(0,1) distribution with 50 data points.
y <- runif(50)
duos_unif <- duos(y, k=4, MH_N=20000)

#Plot the PDF with the data and credible intervals
duos_plot(duos_unif, type="pdf", data=TRUE)

#Plot the CDF with the credible intervals and the empirical CDF
duos_plot(duos_unif, type="cdf", cri=TRUE, data=TRUE)

## --------------------------------------------------------------------------------
## Beta Distribution
## --------------------------------------------------------------------------------

# First run 'duos' on data sampled from a Beta(0.5,0.5) distribution with 300 data points.
y <- rbeta(300, 0.5, 0.5)
duos_arcsin <- duos(y, k=10, MH_N=20000)
#Plot the PDF with the data
duos_plot(duos_arcsin, type="pdf", data=TRUE)

#Plot the CDF with the credible intervals
duos_plot(duos_arcsin, type="cdf", cri=TRUE)

## --------------------------------------------------------------------------------
## Bimodal Distribution
## --------------------------------------------------------------------------------

#Sample 150 random uniforms
U =runif(150)
y = rep(NA,150)
#Sampling from the mixture
for(i in 1:150){
  if(U[i]<.3){
   y[i] = rnorm(1,0,1)
  }else {
   y[i] = rnorm(1,4,1)
  }
}
# First run 'duos' on data sampled from a bimodal distribution with 150 data points.
duos_bimodal <- duos(y, k=6, MH_N=20000)

#Plot the PDF
duos_plot(duos_bimodal)

#Plot the PDF with credible intervals and a histogram of the data
duos_plot(duos_bimodal, cri=TRUE, data=TRUE)

#Plot the CDF with the empirical CDF
duos_plot(duos_bimodal, type="cdf", data=TRUE)

}
