% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_pp.R
\name{duos_pp}
\alias{duos_pp}
\title{Plot the Prior vs. the Posterior}
\usage{
duos_pp(duos_output, parameters = "c", burnin = NA)
}
\arguments{
\item{duos_output}{The list returned by \code{duos} containing the density estimate results.}

\item{parameters}{The group of paramters to plot (see details).}

\item{burnin}{The desired burnin to discard from the results. By default it is half the number of iterations.}
}
\value{
A plot of histograms.
}
\description{
Plots the histograms and densities of simulations from the prior verses the posterior from \code{duos}.
}
\details{
\strong{Options for} \code{parameters}

There are two sets of paramters that can plotted in the histograms: the cut-points and the proportion parameters.
\itemize{
    \item \code{"c"}: Plots the histograms of the cut-points that are ordered and between 0 and 1 (DEFAULT).
    \item \code{"p"}: Plots the histograms of the proportion paramters that sum to one.
  }
}
\examples{

## --------------------------------------------------------------------------------
## Beta Distribution
## --------------------------------------------------------------------------------

# First run 'duos' on data sampled from a beat(2,5) distribution with 50 data points.
y <- rbeta(50, 2, 5)
duos_beta <- duos(y, k=4, MH_N=20000)

#Plot histograms of the priors vs. the posteriors of the cut-point parameters 
duos_pp(duos_beta)

#Plot histograms of the priors vs. the posteriors of the proportion parameters 
duos_pp(duos_beta, parameters = "p")
}
