% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_stat.R
\name{duos_stat}
\alias{duos_stat}
\title{Bayesina Density Estimate of PDF from DUOS.}
\usage{
duos_stat(stat, duos_output, p = NA, burnin = NA, scale = FALSE)
}
\arguments{
\item{duos_output}{The list returned by \code{duos}}

\item{burnin}{The desired burnin to discard from including in the estimate}

\item{scale}{If the data is not already between 0 or 1, this determines if the pdf is returned on the (0,1) transformed data or the original data scale}

\item{x}{A single value or vector of values to calculate the density at}
}
\description{
Calculates an estimate of the density based on the output from \code{duos} for an individual
or vector of values
}
