% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_traceplot.R
\name{duos_traceplot}
\alias{duos_traceplot}
\title{Traceplots from DUOS.}
\usage{
duos_traceplot(duos_output, parameters = "c", plots = "all", burnin = NA,
  start = NA)
}
\arguments{
\item{duos_output}{The list returned by \code{duos}}

\item{burnin}{The desired burnin to discard from including in the estimate}

\item{type}{The type of desired graph (pdf, cdf)}

\item{cr_i}{An option to include credible intervals}

\item{data}{An option to overlay histogram of original data on plot}
}
\description{
Traceplots
}
