% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold.R
\name{gold}
\alias{gold}
\title{Bayesina Density Estimation Using GOLD (Gaussian process On a Log Density).}
\usage{
gold(y, s1, c1, s2, c2, MH_N = 20000, graves = FALSE, scale_l = 1e-05,
  scale_u = 1e-05, poi = NA)
}
\arguments{
\item{y}{A numberic vector to estimate the density on.}

\item{s1}{The standard deviation of the Gaussian prior.}

\item{c1}{The correlation parameter of the Gaussian prior that controls the correlation in the covariance structure.}

\item{s2}{The standard deviation of the Gaussian proposal distribution.}

\item{c2}{The correlation parameter that controls the covariance structure of the Gaussian proposal distribution.}

\item{MH_N}{the number of iterations to run in the algorithm.}

\item{graves}{An option to have the variance of the proposal distribution chosen using an automatice step size selection (insert reference) (DEFAULT is FALSE).}

\item{scale_l}{A value >= 0 controlling the scaling based on the minimum data value (see details).}

\item{scale_u}{A value >= 0 controlling the scaling based on the maximum data value (see details).}

\item{poi}{Points of interest to estimate the density at. \code{gold_pdf} will produce estimates and credible intervals of the density at any point, but \code{mat} returned by \code{gold_pdf} only returns the posterior draws if the density is estimate at that point from the beginning.}
}
\value{
\code{gold} returns a list containing the density estimate results.

\item{\code{G}}{A matrix with the posterior draws for g(x) at a finite number of points. The number of rows is the number of iterations.}
\item{\code{widths}}{A vector with the widths around each each x used for g(x).}
\item{\code{x}}{A vector with the points at which g(x) was estimated.}
\item{\code{y}}{A vector containing the data introduced to \code{gold} for density estimation.}
\item{\code{ar}}{The acceptance rate from the random walk proposals.}
\item{\code{prior}}{A vector of all prior parameter values.}
\item{\code{poi}}{The points of interest added to estimate the density at.}
}
\description{
Estimates a density Bayesian techniques using a Gaussian process on a log density
}
\details{
The density being estimated takes the form below:

\deqn{f(x) =}
\deqn{exp(g(x)) / (\int_{0}^{1} exp(g(u)) du)}

where g(x) is an unknown log density.

Given that g(x) is unknown, the normalizing constant is estimated using a weighted average and the set of unknown paramters that recieve a prior is g(x) at a finite set of points.

This density operates on data between 0 and 1, thus if the input is not between 0 and 1, it is standardized to be between 0 and 1. The formula for scaling is below:

  \deqn{(y-(min(y)-scale_l))/(max(y)+scale_u-(min(y)-scale_l))}.
}
\examples{

## --------------------------------------------------------------------------------
## Beta Distribution
## --------------------------------------------------------------------------------

# First run 'duos' on data sampled from a Beta(5, 1) distribution wiht 200 data points.
# Based on the rule of thumb in the details, 7 cutpoints are used
y <- rbeta(200, 5, 1)
gold_beta <- gold(y = y, s1 = 1, c1 = 0.8, s2 = 0.4, c2 = 0.5, MH_N = 20000)

#Check traceplots
gold_plot(gold_beta, type="pdf", data=TRUE)

#Examine estimate of PDF
gold_plot(gold_beta, type="pdf", data=TRUE)

#Examine estimate of CDF
gold_plot(gold_beta, type="cdf")

#Find probability of being less than 0.4
gold_cdf(c(.4), gold_beta)$cdf
}
