% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_cdf.R
\name{gold_cdf}
\alias{gold_cdf}
\title{Bayesina Density Estimate of CDF from GOLD.}
\usage{
gold_cdf(x, gold_output, burnin = NA, scale = FALSE)
}
\arguments{
\item{x}{A single value or vector of values to calculate the density at}

\item{gold_output}{The list returned by \code{duos}}

\item{burnin}{The desired burnin to discard from including in the estimate}

\item{scale}{If the data is not already between 0 or 1, this determines if the pdf is returned on the (0,1) transformed data or the original data scale}
}
\description{
Calculates an estimate of the CDF based on the output from \code{duos} for an individual
or vector of values
}
