% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_pdf.R
\name{gold_pdf}
\alias{gold_pdf}
\title{Estimate of PDF from GOLD}
\usage{
gold_pdf(x, gold_output, burnin = NA)
}
\arguments{
\item{x}{A single value or vector of values at which to calculate the PDF. These values are to be entered on the scale of the data (i.e. values can fall outside of 0 and 1).}

\item{gold_output}{The list returned by \code{gold} containing the density estimate results.}

\item{burnin}{The desired burnin to discard from the results. If no values is entered, the default is half the number of iterations.}
}
\value{
\code{gold_pdf} returns a list of the PDF results from \code{gold}.

\item{\code{pdf}}{A vector of the posterior mean PDF values at each value in \code{x}.}
\item{\code{cri}}{A matrix with 2 columns and rows equaling the length of \code{x} containing the 95\% credible interval for the PDF at each of the points in \code{x}.}
\item{\code{mat}}{A matrix containing the PDF values for each \code{x} at EACH itertation after the burnin is discarded. The number of columns is the length of \code{x}.}
\item{\code{x}}{A vector containing the values at which to estimate the PDF.}
}
\description{
Calculates a posterior estimate of the PDF based on the output from \code{gold} for an individual
or vector of values.
}
\details{
The function \code{gold_pdf} returns the posterior mean PDF. The PDF is calculated based on the following equation at each iteration:

\deqn{f(x) =}
\deqn{exp(g(x)) / (\int_{0}^{1} exp(g(u)) du)}

where g(x) is an unknown log density.

Given that g(x) is unknown, the normalizing constant is estimated using a weighted average and the set of unknown paramters that recieve a prior is g(x) at a finite set of points.
}
\examples{

## --------------------------------------------------------------------------------
## Beta Distribution
## --------------------------------------------------------------------------------

# First run 'duos' on data sampled from a Beta(2,5) distribution wiht 100 data points.
y <- rbeta(100, 2, 5)
gold_beta <- gold(y, s1 = 1, c1 = 1, s2 = 0.8, c2 = 0.8, MH_N = 20000)
#Calculate pdf at a variety of values
pdf_beta <- gold_pdf(x = c(.01, .25, .6, .9), gold_beta)

#Examine the PDF at 'x'
pdf_beta$pdf

#Examine the credibal intervals of the PDF at 'x'
pdf_beta$cri

## --------------------------------------------------------------------------------
## Normal Distribution
## --------------------------------------------------------------------------------

# First run 'gold' on data sampled from a Normal(0,1) distribution with 200 data points.
y <- rnorm(200, 0, 1)
gold_norm <- gold(y, s1 = 1, c1 = 0.8, s2 = 1, c2 = 0.5, MH_N = 20000)
pdf_norm <- gold_pdf(x = c(-2, -1, 0, 0.8, 1.8), gold_norm)

#Examine the PDF at 'x'
pdf_norm$pdf

#Examine the credibal intervals of the CDF at 'x'
pdf_norm$cri

Histogram of distribution of the PDF density estimate at 0.8
hist(pdf_norm$mat[, 4])
}
