% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_plot.R
\name{gold_plot}
\alias{gold_plot}
\title{Plot Probability Density from GOLD.}
\usage{
gold_plot(gold_output, type = "pdf", burnin = NA, cri = FALSE,
  data = FALSE)
}
\arguments{
\item{gold_output}{The list returned by \code{gold}}

\item{type}{The type of desired graph.}

\item{burnin}{The desired burnin to discard from including in the estimate}

\item{cri}{An option to include credible intervals}

\item{data}{An option to overlay histogram of original data on plot}
}
\description{
Plots the density estimate based on the output from gold.
}
