% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_pp.R
\name{gold_pp}
\alias{gold_pp}
\title{Plot the Prior vs. the Posterior}
\usage{
gold_pp(gold_output, npar = 9, burnin = NA)
}
\arguments{
\item{gold_output}{The list returned by \code{gold} containing the density estimate results.}

\item{npar}{The number of paramters to plot. 3X3 grids of tracepltos are created (DEFAULT is 9).}

\item{burnin}{The desired burnin to discard from the results. By default it is half the number of iterations.}
}
\value{
A plot of histograms.
}
\description{
Plots the histograms and densities of simulations from the prior verses the posterior from \code{gold}.
}
\details{
If \code{npar} is specified to be 9, 9 paramters (as equally spaced as possible) are chosen to plot.
}
\examples{

## --------------------------------------------------------------------------------
## Beta Distribution
## --------------------------------------------------------------------------------

# First run 'duos' on data sampled from a beat(2,5) distribution with 50 data points.
y <- rbeta(50, 2, 5)
duos_beta <- duos(y, k=4, MH_N=20000)

#Plot histograms of the priors vs. the posteriors of the cut-point parameters 
duos_pp(duos_beta)

#Plot histograms of the priors vs. the posteriors of the proportion parameters 
duos_pp(duos_beta, parameters = "p")
}
