% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold_traceplot.R
\name{gold_traceplot}
\alias{gold_traceplot}
\title{Traceplots from DUOS.}
\usage{
gold_traceplot(gold_output, npar = 9, burnin = NA, start = NA)
}
\arguments{
\item{gold_output}{The list returned by \code{duos}}

\item{burnin}{The desired burnin to discard from including in the estimate}

\item{type}{The type of desired graph (pdf, cdf)}

\item{cr_i}{An option to include credible intervals}

\item{data}{An option to overlay histogram of original data on plot}
}
\description{
Traceplots
}
