% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_bed.R
\name{intersect_bed}
\alias{intersect_bed}
\title{Intersect BED}
\usage{
intersect_bed(a, b)
}
\arguments{
\item{a}{First GRanges object}

\item{b}{Second GRanges object}
}
\description{
This function performs an intersection between two GRanges objects
and returns the results in a data.frame
}
\examples{
intersect_bed(a, b)
}
\keyword{BED}
\keyword{GRanges}
