% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{duos}
\alias{duos}
\title{duos}
\usage{
duos(y, MH_N = 20000L, k = 12L, alpha = 1)
}
\arguments{
\item{x}{a vector of strings}
}
\description{
Modify a string in Rcpp.
}
\examples{
duos(x=c('Hello', "C++", 'header', 'files'))

}
