% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duos_pdf.R
\name{duos_pdf}
\alias{duos_pdf}
\title{Estimate of PDF from \code{duos}}
\usage{
duos_pdf(x, duos_output, burnin = NA, scale = FALSE)
}
\arguments{
\item{x}{A single value or vector of values at which to calculate the PDF. These values should be entered on the scale of the data (i.e. values can fall outside of 0 and 1 if data does).}

\item{duos_output}{The list returned by \code{duos} containing the density estimate results.}

\item{burnin}{The desired burnin to discard from the results. If no value is entered, the default is half the number of iterations.}

\item{scale}{This value TRUE/FALSE indicates whether to return scaled or unscaled results IF the original data does not fall between 0 and 1. The default is FALSE (i.e. returns results on the original data scale).}
}
\value{
\code{duos_pdf} returns a list of the PDF results from \code{duos}.

\item{\code{pdf}}{A vector of the posterior mean PDF at each value in \code{x}.}
\item{\code{cri}}{A matrix with 2 columns and a row for each value in \code{x}. It contains the 95\% credible interval for the PDF at each point in \code{x}.}
\item{\code{mat}}{A matrix containing the PDF values for \code{x} at EACH iteration after the burnin is discarded. There is a column for each value in \code{x}.}
\item{\code{x}}{A vector containing the values at which to estimate the PDF.}
}
\description{
Calculates a posterior estimate of the PDF based on the output from \code{duos} for an individual
or vector of values.
}
\details{
The function \code{duos_pdf} returns the posterior mean PDF. The PDF is calculated based on the following equation at each iteration:

\eqn{f(x) =}

\deqn{(\pi_1) / (\gamma_1) , 0 \le x < \gamma_1}
\deqn{(\pi_2) / (\gamma_2-\gamma_1) , \gamma_1 \le x  < \gamma_2}
\deqn{(\pi_3) / (\gamma_3-\gamma_2) , \gamma_2 \le x  < \gamma_3}
\deqn{...  , ... \le x  < ...}
\deqn{(\pi_k) / (\gamma_k-\gamma_{k-1}) , \gamma_{k-1} \le x  < \gamma_k}
\deqn{(\pi_{k+1}) / (1-\gamma_k) , \gamma_k \le x  < 1}

where \eqn{\gamma_1 < \gamma_2 < ... < \gamma_k  is in (0,1) and \pi_1 + \pi_2 + ... + \pi_{k+1} = 1}
}
\examples{
## --------------------------------------------------------------------------------
## Uniform Distribution
## --------------------------------------------------------------------------------

# First run \\code{duos} on data sampled from a Unif(0,1) distribution with 70 data points.
y <- runif(70)
duos_unif <- duos(y = y, k = 4, MH_N = 20000)

# Estimate the PDF at a vector of values
pdf_unif <- duos_pdf(x = c(.1, .5, .65, .98), duos_unif)

# Examine the PDF at \\code{x}
pdf_unif$pdf

# Examine the credible intervals of the PDF at \\code{x}
pdf_unif$cri

## --------------------------------------------------------------------------------
## Gamma Distribution
## --------------------------------------------------------------------------------

# First run \\code{duos} on data sampled from a Gamma(2,2) distribution with 90 data points.
y <- rgamma(90, 2, 2)
duos_gamma <- duos(y = y, MH_N = 20000)

# Estimate the PDF at a vector of values
pdf_gamma <- duos_pdf(x = c(0.4, 1, 2, 3), duos_gamma)

# Examine the PDF at \\code{x}
pdf_gamma$pdf

# Examine the credible intervals of the PDF at \\code{x}
pdf_gamma$cri

#Plot a histogram of distribution of the posterior draws for the PDF estimate at 1
hist(pdf_gamma$mat[,2])

}
