% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold.R
\name{gold}
\alias{gold}
\title{Bayesina Density Estimation Using GOLD (Gaussian process On a Log Density).}
\usage{
gold(y, s1, c1, s2, c2, MH_N)
}
\arguments{
\item{y}{a numeric vector containing the data on which to estimate a density}

\item{s1}{the standard deviation of the Gaussian prior}

\item{c1}{the correlation parameter of the Gaussian prior that controls the correlation in the covariance structure}

\item{s2}{the standard deviation of the Gaussian proposal distribution}

\item{c2}{the correlation paramtet that controls the covariance structure of the Gaussian proposal distribution}

\item{MH_N}{the number of iterations}
}
\description{
Bayesian density estimation through a Gaussian process on a log density
}
